mappers? :(

+++ src/main/java/com/example/workout_tracker/dtos/WorkoutEntryDto.java
@@ -2,6 +2,8 @@ package com.example.workout_tracker.dtos;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class WorkoutEntryDto {
     private Long id;
@@ -9,5 +11,5 @@ public class WorkoutEntryDto {
     private String description;
     private int sets;
     private int reps;
-    private double weight;
+    private BigDecimal weight;
 }
